weddingrestaurantapi/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cloudinary==1.41.0
cryptography==43.0.0
Django==5.1
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-oauth-toolkit==2.4.0
djangorestframework==3.15.2
idna==3.7
jwcrypto==1.5.6
oauthlib==3.2.2
pillow==10.4.0
pycparser==2.22
PyMySQL==1.1.1
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
